# Remove eager loding of shipping_methods and shipping_rates from customer_totals so that report runs faster