#### What? Why?

Closes #4547

These includes make the main query much slower.
Without these includes I see 278 queries instead of 209 queries but the report loads in 50secs instead of 4.48mins.

These are my local tests with different includes in CustomerTotals#line_item_includes:
```
        # original version: https://github.com/openfoodfoundation/openfoodnetwork/commit/8a57977b448937b81a2f058546183e8d269d707a
        [{ variant: :product,
           order: [:bill_address, :shipments, :order_cycle, :adjustments] }] # 42secs - 649 queries

        [{ variant: :product,
           order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                                        :shipments] }] # 42s - 564 queries

        [{ variant: :product,
           order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                   :user, :distributor, :shipments] }] # 44s 516 queries

        [{ variant: [{ option_values: :option_type }, { product: :supplier }],
           order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                   :user, :distributor, :shipments] }] # 50 secs 278 queries

        [{ variant: :product,
           order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                   shipments: { shipping_rates: :shipping_method }] }] # 1,92mins - 495 queries

        # LIVE version
        [{ variant: [{ option_values: :option_type }, { product: :supplier }],
           order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                   :user, :distributor, shipments: { shipping_rates: :shipping_method }] }] # 4.48mins - 209 queries
``` 

#### What should we test?
Run the report with live data and see that is runs much faster without timing out.
This wont break the report for sure so we may just push this to prod and see the impact.

#### Release notes
Changelog Category: Changed
Improved performance of Orders and Fullfilment report Customer Totals by improving the database query.