@luisramos0 2019-12-10
I spent a full day replacing the very long IN clause from the query (100k order ids and 70k line items ids) in #4549 just to realize that the main problem was in the includes...
The report with this PR and 4549 runs in 42sec (275 queries) instead of 50secs (279 queries), so 4549 will also improve performance.