@sauloperez 2019-12-11
I don't really understand how this improves the situation. Two questions come to my mind:

* If we need to remove these, did we introduce a regression at some point? Adding includes is something we just did recently.
* If the data from these tables is needed in the report, does this mean we now introduce N+1? How can that be faster?