@luisramos0 2019-12-11
Adding includes can mess up the execution plan of a query (imagine for example joining all orders in the DB with their associated tables before filtering the list of orders to the orders in the OC). So in this case removing an include makes the query go faster in the DB and then we load the data in N+1 queries, in this case, instead of 209 we make 278 queries. We run 69 more queries but the initial very complex query runs in a few seconds instead of 5 minutes.