@luisramos0 2019-12-11
I just confirmed the load times locally, with uk live data and the exact same criteria as the S2 bug, I get 5mins before this PR and 52secs after this PR.