@kristinalim 2019-12-11
Thanks for your tip a couple of weeks ago @luisramos0 about swap just not being freed. About current used swap in UK, I ran `vmstat 1` and noting that the used swap seems to be inactive:

```
procs -----------memory---------- ---swap-- -----io---- -system-- ------cpu-----
 r  b   swpd   free   buff  cache   si   so    bi    bo   in   cs us sy id wa st
 1  0 549580 1602588 312908 1649996    0    0     0    72  166  322  0  0 100  0  0
 0  0 549580 1602588 312908 1649996    0    0     0    52  138  273  0  0 100  0  0
 0  0 549580 1602652 312908 1649996    0    0     0     4  613 1199  2  0 98  0  0
```