@luisramos0 2019-12-12
nice @kristinalim 
I wonder how do you see that the used swap seems to be inactive. I know this is a common thing from some past experience but I cant remember any details :rofl: 

Anyway I will test this afternoon again and see if when running the report in master I see the swap being used. Could be it.

I am running with uk live data, all the DB without any anonymization or truncation.