@lin-d-hop 2019-12-12
I gave this a gander on all O&F reports.

Everything looks fine.

I seem to be locked out of all accounts bulk data for testing.... On UK and FR staging. So no tests on performance have been done at all. That can happen on production.

Good to go :-)